+++ go.mod
@@ -3,8 +3,11 @@ module helm-edit
 go 1.14
 
 require (
+	github.com/Masterminds/semver v1.5.0 // indirect
+	github.com/Masterminds/sprig v2.22.0+incompatible // indirect
 	github.com/spf13/cobra v1.0.0
 	helm.sh/helm/v3 v3.2.1
+	k8s.io/helm v2.16.7+incompatible
 	rsc.io/letsencrypt v0.0.3 // indirect
 	sigs.k8s.io/yaml v1.2.0
 )
