+++ install.sh
@@ -1,8 +1,13 @@
 #!/usr/bin/env bash
 
 set -euf -o pipefail
-
-HELM_EDIT_VERSION=${HELM_EDIT_VERSION:-"0.3.0"}
+DEFAULT_VERSION="0.4.0"
+HELM_VERSION=$(helm version --short | head -c2)
+if [[ ${HELM_VERSION} -eq "v2" ]]; then
+  echo "This plugin version support only Helm 3. Defaulting to previous version"
+  DEFAULT_VERSION="0.3.0"
+fi
+HELM_EDIT_VERSION=${HELM_EDIT_VERSION:-"${DEFAULT_VERSION}"}
 
 file="${HELM_PLUGIN_DIR:-"$(helm home)/plugins/helm-edit"}/helm-edit"
 os=$(uname -s | tr '[:upper:]' '[:lower:]')
