+++ plugin.yaml
@@ -1,5 +1,5 @@
 name: "edit"
-version: "0.3.0"
+version: "0.4.0"
 usage: "edit a release"
 description: |-
   Edit a release.
