--- test/test/foo/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-appVersion: "1.0"
-description: A Helm chart for Kubernetes
-name: foo
-version: 0.1.0
