--- test/test/foo/templates/service.yaml
@@ -1,19 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "foo.fullname" . }}
-  labels:
-    app.kubernetes.io/name: {{ include "foo.name" . }}
-    helm.sh/chart: {{ include "foo.chart" . }}
-    app.kubernetes.io/instance: {{ .Release.Name }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
-spec:
-  type: {{ .Values.service.type }}
-  ports:
-    - port: {{ .Values.service.port }}
-      targetPort: http
-      protocol: TCP
-      name: http
-  selector:
-    app.kubernetes.io/name: {{ include "foo.name" . }}
-    app.kubernetes.io/instance: {{ .Release.Name }}
