--- test/test/foo/templates/tests/test-connection.yaml
@@ -1,18 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: "{{ include "foo.fullname" . }}-test-connection"
-  labels:
-    app.kubernetes.io/name: {{ include "foo.name" . }}
-    helm.sh/chart: {{ include "foo.chart" . }}
-    app.kubernetes.io/instance: {{ .Release.Name }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
-  annotations:
-    "helm.sh/hook": test-success
-spec:
-  containers:
-    - name: wget
-      image: busybox
-      command: ['wget']
-      args:  ['{{ include "foo.fullname" . }}:{{ .Values.service.port }}']
-  restartPolicy: Never
